+++ www/CHANGES.md
@@ -5,7 +5,7 @@ A Little Big Adventure 2 engine remake change log file
 
 ## Unreleased
 ### Added
-* 
+* Raw sprites from SPRIRAW.HQR file (eg. Magicball Twinsen's Secret Room)
 
 ### Fixed
 *